+++ Editor/Scripts/ShaderGraphGUI.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-#if USING_URP || USING_HDRP
+#if USING_URP || USING_HDRP || ( UNITY_SHADER_GRAPH_12_OR_NEWER && !GLTFAST_FORCE_BUILTIN_SHADERS)
 #define GLTFAST_SHADER_GRAPH
 #endif